refactor: move venue components to venue folder so that they are next to the venuesModule that uses them
