+++ src/features/home/components/modules/venues/VenueTile.tsx
@@ -5,7 +5,7 @@ import { useTheme } from 'styled-components'
 import styled from 'styled-components/native'
 
 import { VenueResponse } from 'api/gen'
-import { VenueDetails } from 'features/home/components/VenueDetails'
+import { VenueDetails } from 'features/home/components/modules/venues/VenueDetails'
 import { VenueTypeLocationIcon } from 'features/home/components/VenueTypeLocationIcon'
 import { analytics } from 'libs/firebase/analytics'
 import { GeoCoordinates } from 'libs/geolocation'
